Safe enable lastfm (#4)
* Load Last.FM keys from an external location if they exist

This way we can easily load up someones API keys without risking
adding them to the repository. Just put a file with the correct
name at any location above your checkout and it will be loaded
instead of the file in git.

* Enable lastfm only if there's a non-empty API key

If we have an actual API key then we will show the settings
icon. Otherwise leave it hidden.